pyproject.toml
[tool.poetry]
name = "exporter"
version = "0.1.0"
description = ""
authors = ["Vlad Nedelcu <nedelcuvd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.5"
Jinja2 = "^3.1.2"
pdfkit = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
